+++ libraries/test_utils/src/main/java/androidx/media3/test/utils/VideoFrameProcessorTestRunner.java
@@ -38,7 +38,7 @@ import androidx.annotation.Nullable;
 import androidx.media3.common.ColorInfo;
 import androidx.media3.common.DebugViewProvider;
 import androidx.media3.common.Effect;
-import androidx.media3.common.FrameInfo;
+import androidx.media3.common.Format;
 import androidx.media3.common.GlTextureInfo;
 import androidx.media3.common.SurfaceInfo;
 import androidx.media3.common.VideoFrameProcessingException;
@@ -283,8 +283,8 @@ public final class VideoFrameProcessorTestRunner {
               @Override
               public void onInputStreamRegistered(
                   @VideoFrameProcessor.InputType int inputType,
-                  List<Effect> effects,
-                  FrameInfo frameInfo) {
+                  Format format,
+                  List<Effect> effects) {
                 videoFrameProcessorReadyCondition.open();
               }
 
@@ -338,13 +338,14 @@ public final class VideoFrameProcessorTestRunner {
             @Nullable ColorInfo colorInfo = MediaFormatUtil.getColorInfo(mediaFormat);
             videoFrameProcessor.registerInputStream(
                 INPUT_TYPE_SURFACE,
-                effects,
-                new FrameInfo.Builder(
-                        colorInfo == null ? ColorInfo.SDR_BT709_LIMITED : colorInfo,
-                        mediaFormat.getInteger(MediaFormat.KEY_WIDTH),
-                        mediaFormat.getInteger(MediaFormat.KEY_HEIGHT))
+                new Format.Builder()
+                    .setColorInfo(colorInfo == null ? ColorInfo.SDR_BT709_LIMITED : colorInfo)
+                    .setWidth(mediaFormat.getInteger(MediaFormat.KEY_WIDTH))
+                    .setHeight(mediaFormat.getInteger(MediaFormat.KEY_HEIGHT))
                     .setPixelWidthHeightRatio(pixelWidthHeightRatio)
-                    .build());
+                    .build(),
+                effects,
+                /* offsetToAddUs= */ 0);
             try {
               awaitVideoFrameProcessorReady();
             } catch (VideoFrameProcessingException e) {
@@ -374,11 +375,14 @@ public final class VideoFrameProcessorTestRunner {
     videoFrameProcessorReadyCondition.close();
     videoFrameProcessor.registerInputStream(
         INPUT_TYPE_BITMAP,
-        effects,
-        new FrameInfo.Builder(colorInfo, inputBitmap.getWidth(), inputBitmap.getHeight())
+        new Format.Builder()
+            .setColorInfo(colorInfo)
+            .setWidth(inputBitmap.getWidth())
+            .setHeight(inputBitmap.getHeight())
             .setPixelWidthHeightRatio(pixelWidthHeightRatio)
-            .setOffsetToAddUs(offsetToAddUs)
-            .build());
+            .build(),
+        effects,
+        offsetToAddUs);
     awaitVideoFrameProcessorReady();
     checkState(
         videoFrameProcessor.queueInputBitmap(
@@ -396,10 +400,14 @@ public final class VideoFrameProcessorTestRunner {
     videoFrameProcessorReadyCondition.close();
     videoFrameProcessor.registerInputStream(
         INPUT_TYPE_BITMAP,
-        effects,
-        new FrameInfo.Builder(colorInfo, width, height)
+        new Format.Builder()
+            .setColorInfo(colorInfo)
+            .setWidth(width)
+            .setHeight(height)
             .setPixelWidthHeightRatio(pixelWidthHeightRatio)
-            .build());
+            .build(),
+        effects,
+        /* offsetToAddUs= */ 0);
     awaitVideoFrameProcessorReady();
     for (Pair<Bitmap, TimestampIterator> frame : frames) {
       videoFrameProcessor.queueInputBitmap(frame.first, frame.second);
@@ -410,10 +418,14 @@ public final class VideoFrameProcessorTestRunner {
       throws VideoFrameProcessingException {
     videoFrameProcessor.registerInputStream(
         INPUT_TYPE_TEXTURE_ID,
-        effects,
-        new FrameInfo.Builder(colorInfo, inputTexture.width, inputTexture.height)
+        new Format.Builder()
+            .setColorInfo(colorInfo)
+            .setWidth(inputTexture.width)
+            .setHeight(inputTexture.height)
             .setPixelWidthHeightRatio(pixelWidthHeightRatio)
-            .build());
+            .build(),
+        effects,
+        /* offsetToAddUs= */ 0);
     videoFrameProcessor.setOnInputFrameProcessedListener(
         (texId, syncObject) -> {
           try {